+++ weldx/tests/test_utility.py
@@ -641,3 +641,11 @@ class TestWeldxExampleCompareNested(unittest.TestCase):
 
         assert csm_copy != csm_org
         assert not ut.compare_nested(self.a, self.b)
+
+
+def test_is_interactive():
+    """Assert that the Pytest session is not recognized as interactive.
+
+    >>> assert not ut.is_interactive_session()
+    """
+    assert not ut.is_interactive_session()